requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk==3.5.0
gunicorn
langchain>=0.0.308
openai
yfinance
pydantic
tiktoken
youtube_search
googlesearch-python